feat(organization):reorganize the file structure
