+++ backend/src/controllers/cloudinaryUpload/removeImage.js
