+++ frontend/src/question.txt
@@ -1,7 +1,7 @@
 server file:
 import express from 'express';
-import databaseConnection from './config/db.js';
-import router from './routes/route.js';
+import databaseConnection from ./src/config/db.js';
+import router from './src/routes/route.js';
 import cors from 'cors';
 import cookieParser from 'cookie-parser';
