+++ frontend/tsconfig.json
@@ -19,6 +19,6 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["src"],
+  "include": ["src", "utils"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
